The code is working!
The Paperclip evolution code is working, with just 2 of 4 algorithms up and running: top 10 always survive, and random mutation using tournament selection. The user inputs the number of segments of the antenna and the number of generations. Current fitness score attempts to maximize the curlyness of the antenna about the z direction